fix: move schema
